update asic_demo test to check for 0 tns for setup and hold
